Place sprites in the game world
Actor sprites (so far only the player) are placed in the game world.
Fixed minor bug in loading the game tile data.
Added positional information to the actor sprites.
Added an immediate update to the universe to allow for pre-calculating
positional data before the scene is rendered for the first time.